Functionality update.
Added functionality that automatically centered slideshows vertically
when the user interacts, so that it doesn’t end up offscreen. Will
eventually add a toggle for it.

Also added page loader functionality (with optional versions of
loaders, which will grow with time). For those impatient users that
want to know things are loading when they click a link.